继续修复大量和typing有关的bug

+++ linpg/battle/weather.py
@@ -21,7 +21,7 @@ class WeatherSystem:
                     imgId=get_random_int(0, len(self.__img_list) - 1),
                     size=get_random_int(5, 10),
                     speed=get_random_int(1, 4),
-                    x=get_random_int(1, Display.get_width() * 1.5),
+                    x=get_random_int(1, int(Display.get_width() * 1.5)),
                     y=get_random_int(1, Display.get_height()),
                 )
                 for i in range(entityNum)
